+++ src/main/kotlin/be/hize/afknotifier/events/ChatEvent.kt
@@ -1,0 +1,5 @@
+package be.hize.afknotifier.events
+
+import net.minecraft.util.IChatComponent
+
+class ChatEvent(val message: String, val chatComponent: IChatComponent,) : ModEvent()