+++ src/main/kotlin/be/hize/afknotifier/utils/NumberUtil.kt
@@ -1,0 +1,7 @@
+package be.hize.afknotifier.utils
+
+import java.text.NumberFormat
+
+object NumberUtil {
+    fun Number.addSeparators() = NumberFormat.getNumberInstance().format(this)
+}