+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change Log
 
+## 1.0.6
+* Added an option to send message only when island change is caused by a server restart (Game update/Scheduled reboot)
+* Added an option to send a message when a reboot is detected (When it appears in the chat)
+* You can add %%island%% in the IS Leave Message config message to replace it with the configured island.
+
 ## 1.0.5
 * Fixed the HTTP request not working for some people.